ElmyraService: Fix double vibration when device is in deep sleep
When the device's AP is in deep sleep (i.e. suspended), SLPI takes over
the vibration driver IC and handles Elmyra vibrations by itself without
waking the AP up. Only once the gesture has been detected does it send a
CHRE event to wake the AP for gesture handling.

It turns out that this is what the hostSuspended and hapticConsumed
properties in the GestureDetected message mean, so let's document it in
the protobuf descriptor while we're at it.

Signed-off-by: spezi77 <spezi7713@gmx.net>